ch11/requirements.txt
langgraph==0.5.1
langchain==0.3.26
langchain_openai==0.3.27
langchain_community==0.3.27
chromadb==1.0.15
aiohttp==3.12.13
tiktoken==0.9.0
beautifulsoup4==4.12.4
html2text==2025.4.15
python-dotenv==1.1.1
langgraph-supervisor==0.0.27
fastmcp==2.10.5
langchain-mcp-adapters==0.1.9